Setup for publishing to npm
Summary:
A bunch of changes to support publishing Xrpa to npm. Primary change is that the package names all now contain the `xrpa/` scope. Also modified the publish script to keep the version numbers in sync with each other.

Bumped version number to 1.3.0 and updated all modules to the oss directory.

Reviewed By: samueltfb

Differential Revision: D59664019

fbshipit-source-id: 2f36431656faa6de9fcdee99b8c7f963a5d4e973