correctly handle Unity editor changes to fields
Summary: When a field is modified in Unity, the change bypasses the property setter function, so the change never propagates anywhere. This diff adds an OnValidate() function which Unity calls after a field is changed in the editor. Unfortunately it does not tell us which fields changed, so in the normal case the generated code has to set every field dirty. In the synthetic object case, however, the generated code manually checks if the value is different.

Reviewed By: samueltfb

Differential Revision: D59826133

fbshipit-source-id: 98c231810df716a36eb7355098d422e3ae60e091