Fix for synthetic object component parameters not being passed to the effect when calling Run()
Summary: An effect in the Synecdoche project was hitting this case where changing the effect parameters was having no effect. I had not run into this previously because typically I use AutoRun with longer running effects, and the parameters were getting set correctly after the effect is running. They were already working correctly for the Spawn() function too, it is just Run() that was missing the parameter setter calls.

Reviewed By: samueltfb

Differential Revision: D59922225

fbshipit-source-id: f5085703613da5e0e3016d528885bdcdaf15f4bc