Add option file to enable unsafe mode just for the Xrpa Unity runtime
Summary: Previously users of Xrpa had to enable unsafe mode for their entire project (in Player settings). Really only the Xrpa package needs it enabled, and this option file added to the package accomplishes that. It is also carries over to the generated Visual Studio project, unlike the project setting, so VS no longer complains.

Reviewed By: samueltfb

Differential Revision: D59974668

fbshipit-source-id: 993d499cbeab7babb2d50b5c36682b866b6c7013