Fix for XredOutput crashing with pulsar dongle plugged in
Summary: The pulsar dongle exposes itself as a COM UART as well as a HID device, and unfortunately it uses the same PID and VID as the HDK dongle. This was causing the signal output code to crash. Rearranged some code so that we try to query the device info before registering it as a valid dongle, so that only BuzzDuino and HDK-1 will be used.

Reviewed By: samueltfb

Differential Revision: D63336437

fbshipit-source-id: 663a4f75ef41cbb312b1612877bcaa9f3edc66fe