Use camera calibration file, add setting for code size
Summary:
Added a command line setting to configure the printed QR code size for tracking. Defaults to 19cm, which is what I measured my printout on 8.5x11 paper to be.

Added loading of the camera calibration file (supplied by Ocean in their demos folder), and pulling that camera information in to the tracker.

Added an attempt to use the C920 camera by name before falling back to webcam 0, in case there are more than one camera connected.

Reviewed By: samueltfb

Differential Revision: D64441082

fbshipit-source-id: 8497bd2c84d153a1ace1d26755b052afc77d26c4