Setup a postinstall script that runs the codegen
Summary:
A common mistake seems to be updating the dependency versions, running `yarn`, but then forgetting to also run `yarn update`. Also, perhaps `yarn update` is not an obvious name.

Added a postinstall script to the xrpa-init package.json so that the codegen runs after yarn install. Also added an alias `yarn codegen` for `yarn update`.

Reviewed By: samueltfb

Differential Revision: D64910416

fbshipit-source-id: a4fbc881f52f8664a5cdc755a2dddfe77502ca3b