Make unknown speaker result consistent
Summary:
The speaker_id gets filled with empty string as the default, so it should be set to empty string when it cannot identify the speaker. This makes the logic for using the module simpler.

I can see an argument for wanting to know the difference between "nobody speaking" and "unknown speaker", but I think that will be better coming from moving VAD out and having speech-events as a structured output.

Reviewed By: archthegit

Differential Revision: D84717757

fbshipit-source-id: 091e266da4f0933d6fe6a4c8a2f497e77214b600