Print error if receiving the incorrect sample rate
Summary: I don't want to spend the CPU cycles on resampling the streaming inbound audio, so at the moment I just changed it so it prints out an error if the inbound sample rate is not correct. It is much better to just open the microphone audio stream at 16khz.

Reviewed By: archthegit

Differential Revision: D84717918

fbshipit-source-id: 00c45eabbcedbe4452075fd175710de34be31646