Build Module script and corrected Sandcastle command
Summary:
The previous base sandcastle command was inheriting from the wrong parent, hadn't noticed it was extending from a command meant for electron. I've replaced it with a generic lego command for signing with autograph

Updated the buildModule script to be generic.

Once this merges, need to whitelist the commands to try it out

Reviewed By: conordickinson

Differential Revision: D85967122

fbshipit-source-id: 8c15f3851ccffdc6990f259d9dff3a53032f8183