Handle failed initializeMemory call and improve version mismatch handling
Summary:
I noticed while doing the heartbeating that a version or schema mismatch was not actually being handled, which could (and likely did) cause corruption of the transport memory. Now handling that by releasing the memory which will cause all locks to fail.

While in there, I also pulled the version and schema comparison logic up into initializeMemory() so that it can be smarter about how it handles those failuers. It tries to recover by re-initializing whenever it can, but if the memory page is still actively used and the version/schema does not match, then it errors out.

Reviewed By: archthegit

Differential Revision: D84844994

fbshipit-source-id: 5423c7b38bd9411d93596ead0def6f61f394ac9e