Add a transport timeout and heartbeating
Summary:
This change is to deal with the fact that shared memory files never get cleaned up on Mac. Added an update timestamp to the transport header, which gets set every time the mutex gets locked. The reconciler now makes sure to lock it if the timestamp is over 1 second old.

When the memory page is acquired, the code checks the age of the timestamp, and if it is over a 20s old it assumes it is stale and re-initializes.

Reviewed By: archthegit

Differential Revision: D84830523

fbshipit-source-id: c171044dac981bc4e3756793e27411590b34e121