Create standard image type and conversion util for Python
Summary:
Made a standard Python image type like we have in C++ and C# already.

Combined all the various image->pil conversion logic into a standard conversion function in the runtime utils. Also set the flag to allow Pillow to load truncated jpegs (which seemed to only be an issue in Python 3.11, not 3.12). This fixes the various warnings in the Gesture module.

Reviewed By: archthegit

Differential Revision: D86562929

fbshipit-source-id: b362b77bbf942e4e9b373e58714ca6017d11caea