Fix for heartbeat age computation mismatch metween C++ and other languages
Summary:
Fixes a bug where the Python and C# implementations of the transport update age calculation were treating the stored value as microseconds, while C++ was treating it as milliseconds. No idea how that got messes up, but this fixes it.

Also fixes an integer rounding bug in C# that was making it miscompute the timestamp in nanoseconds.

Reviewed By: archthegit

Differential Revision: D86785836

fbshipit-source-id: ece90c46852ee6c1ec81a4318d4079ccf3c952cf