Use the transport version and schema hash in shared memory file name
Summary:
On Mac, shared memory is implemented via memory-mapped files, and trying to open one with a different size than previously fails. Solving this problem by embedding the transport version and schema hash in the shared memory file name. This makes the runtime version checks a little redundant, but they are still useful for validation and in case of the rare hash collision (since the file name is only using 32 bits of the hash).

Note that this changes the shared memory name (and mutex name) used in Windows as well, but I'm ok with that. Should not cause any problems.

Since the shared memory files are in /tmp, the old ones should be cleaned up automatically by the OS.

Reviewed By: archthegit

Differential Revision: D86789379

fbshipit-source-id: 207c8770a53a9292416bee86b9e2fac22795827a