Python dataflow support
Summary:
Adds support for generating dataflow programs in to Python. This was mostly an exercise in making the dataflow codegen fully use the generic codegen interface, and I expect there are still some missing gaps here that I did not find (it is hard to fully exercise the dataflow codegen code paths).

While inspecting the generated output, I noticed that there was an existing bug around string replacement, where it would leave in an extra closing curly brace during parameter replacement. Fixed that for all targets.

Reviewed By: archthegit

Differential Revision: D87099037

fbshipit-source-id: 9a3887cd9a095d6de9421a13f915f6b0562821de