MacOS mutex lock fix for Python and C#
Summary:
Need to keep the file descriptor alive during the lifetime of the mutex lock, or else flock will immediately release the file lock (and thus the mutex).

Also removed the sleep from within the lock retry, because it basically made it never retry (our timeout is 1ms and the retry sleep was 1ms, so timeout was hit after one sleep). Now it busy-retries, which reduces lock failures.

Reviewed By: archthegit

Differential Revision: D88325061

fbshipit-source-id: c39cb548b4ef81c4ba5b75c3e94fe73410f5a889