C++ optimization
Summary:
Found some performance abnomalities between languages when I was running fuzz testing for large byte arrays. This diff addresses the C++ read performance.

Resizing a std::vector<uint8_t> causes it to initialize the memory in a fairly slow way (setting one byte at a time, I think), which is a problem when all you want to do is copy memory into it. There doesn't seem to be a good way around that behavior, even with a custom allocator class (I tried several options). So instead I switched the MemoryAccessor to read/write with a custom ByteVector class which does not initialize the memory.

Speeds up image-size reads by 2 orders of magnitude (!!).

This change required me to move ImageTypes.h into the Xrpa runtime (which matches Python and C#), which had a domino effect on the modules. Also moved the ImageConversion code into the one module that actually uses it (Camera) and removed the old ImageDebugWindow class that has been unused since porting the ImageViewer to Python.

Reviewed By: archthegit

Differential Revision: D88655962

fbshipit-source-id: d86f1ef9ea91457acd777736fb1e1a09de4be9d4