added build:release top level script, build fixes
Summary:
Adds a build:release top-level script which runs (the poorly named) oss-publish script on each package.

Changed the yarn-workspaces script to check that a script is valid in a package before running it.

Removed the invalid oss-publish script from xred-livestack.

Added /bin/ directory to gitignore files that were missing it.

Reviewed By: archthegit

Differential Revision: D88780038

fbshipit-source-id: c11077c9cecc946970d6228b628472490430d8d2