Lock-free interprocess SPMC ring buffer
Summary:
This is an implementation of a single-producer multiple-consumer ring buffer that can work across processes, across languages. Uses appropriate atomic operations with memory fences to make sure reads and writes are in the correct order.

For Python it uses a small C extension to expose the atomic intrinsics. Somewhat annoying but there doesn't seem to be another way to accomplish this in Python.

Reviewed By: samueltfb

Differential Revision: D88547102

fbshipit-source-id: cfcf87543a916715d3b517d0d34c1504536ac4c0